// src/interfaces/search.interface.ts
export interface FoodTruckResponse {
  locationId: string;
  applicant: string;
  address: string;
  locationDescription: string;
  blocklot: string;
  permit: string;
  status: string;
  approved: string;
  schedule: string;
  daysHours: string;
  foodItems: string;
  distance: string;
}

export interface ResponseKeys {
  Latitude: string;
  Longitude: string;
  locationid: string;
  Applicant: string;
  Address: string;
  LocationDescription: string;
  blocklot: string;
  permit: string;
  Status: string;
  Approved: string;
  Schedule: string;
  dayshours: string;
  FoodItems: string;
}